requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
sentence-transformers>=2.2.0
chromadb>=0.4.0
rank-bm25>=0.2.2
python-docx>=1.1.0
openai>=1.0.0
python-multipart>=0.0.6
